XML Note Tools: trunk <- working copy
!!This commit is only to store files - the code here still has many bugs!!!

General
This represents an intermediate solution to a major revision of the code. The goal is to separate pieces for the Tools (user accessible functions), Module (inaccessible support functions), and Panel (GUI input functions). Code in the Tools should be sparse, static, and user accessible. Code in the Module should be hidden and may be continually optimized (ie via XOPs). Code in the Panel should be hidden and may evolve.